Switch to quint16 tokens for 8bit morse
Allows us to use full 8bit morse symbols with the 9th bit as flag.  Table size goes from 256 to 512 as a result

+++ plugins/MorseDigitalModem/MorseIQMaker/morsegen.h
@@ -26,7 +26,7 @@ public:
 	quint32 genChar(CPX *out);
 	quint32 genWord(CPX *out);
 	//For testing, dump whole table
-	quint32 genToken(CPX *out, quint8 token);
+	quint32 genToken(CPX *out, quint16 token);
 	quint32 genTable(CPX* out, quint32 index);
 
 private:
